[IMP] account_brand: compute account_id at move_line creation
This module has changed a lot between Odoo versions, some features got lost on the way.
The account should be computed when a new move line is added, not only when the onchange triggers on the move.
The onchange on the move should trigger if the brand or the company changes.