[FIX] contract_brand: combine analytic distribution defined by analytic distribution models with analytic on brands
Before this commit the analytic distribution was always erased on contract lines: if the contract is linked to a brand,
only the analytic of the brand was considered. If the contract wasn't linked to a brand, the analytic was erased.

After this commit: we consider analytic distribution models, to define the analytic distribution. And we combine
it with the analytic distribution of the brand, if needed